create login and register sample flow

+++ apps/api/src/main/java/com/hbloc/youtube_tool_demo/user/infrastructure/UserStatusRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserStatusRepository extends JpaRepository<UserStatus, Long> {
+public interface UserStatusRepository extends JpaRepository<UserStatus, Integer> {
     Optional<UserStatus> findByCode(String code);
 }
